--- a/pyproject.toml
+++ b/pyproject.toml
@@ -64,6 +64,9 @@ norwegianblue = "norwegianblue.cli:main"
 [tool.hatch]
 version.source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/norwegianblue/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"