With Python 3.12.4:

```sh
python -X importtime -c "import norwegianblue" 2> import.log && tuna import.log
```

Cuts out the 22ms `importlib.metadata` import.

# `main`

<img width="1624" alt="image" src="https://github.com/user-attachments/assets/494b7348-633f-4b17-9cc0-34c7d609b9c4">

```console
❯ hyperfine -w 8 "eol --version"
Benchmark 1: eol --version
  Time (mean ± σ):      55.1 ms ±   3.8 ms    [User: 43.3 ms, System: 10.4 ms]
  Range (min … max):    50.9 ms …  71.4 ms    53 runs
```

# PR

<img width="1624" alt="image" src="https://github.com/user-attachments/assets/fddd61d3-3d58-491d-8ad4-8d19525201d5">

```console
❯ hyperfine -w 8 "eol --version"
Benchmark 1: eol --version
  Time (mean ± σ):      38.2 ms ±   1.7 ms    [User: 29.6 ms, System: 7.3 ms]
  Range (min … max):    35.3 ms …  41.9 ms    73 runs
```